Refactored CLI commands
