+++ cmd/kbc/main.go
@@ -3,15 +3,15 @@ package main
 import (
 	"os"
 
-	"github.com/keboola/keboola-as-code/internal/pkg/cli"
+	"github.com/keboola/keboola-as-code/internal/pkg/cli/cmd"
+	interactivePrompt "github.com/keboola/keboola-as-code/internal/pkg/cli/prompt/interactive"
 	"github.com/keboola/keboola-as-code/internal/pkg/env"
 	"github.com/keboola/keboola-as-code/internal/pkg/filesystem/aferofs"
-	"github.com/keboola/keboola-as-code/internal/pkg/interaction"
 )
 
 func main() {
 	// Interactive prompt
-	prompt := interaction.NewPrompt(os.Stdin, os.Stdout, os.Stderr)
+	prompt := interactivePrompt.New(os.Stdin, os.Stdout, os.Stderr)
 
 	// Load Os Envs
 	osEnvs, err := env.FromOs()
@@ -20,6 +20,6 @@ func main() {
 	}
 
 	// Run command
-	cmd := cli.NewRootCommand(os.Stdin, os.Stdout, os.Stderr, prompt, osEnvs, aferofs.NewLocalFsFindProjectDir)
-	os.Exit(cmd.Execute())
+	rootCmd := cmd.NewRootCommand(os.Stdin, os.Stdout, os.Stderr, prompt, osEnvs, aferofs.NewLocalFsFindProjectDir)
+	os.Exit(rootCmd.Execute())
 }
